+++ src/main/java/io/github/astrapi69/modify/ModifyFileExtensions.java
@@ -39,7 +39,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.function.BiFunction;
 
-import de.alpharogroup.collections.list.ListFactory;
+import io.github.astrapi69.collections.list.ListFactory;
 import io.github.astrapi69.modify.api.FileChangeable;
 
 /**