+++ src/main/java/io/github/astrapi69/search/FileSearchExtensions.java
@@ -30,7 +30,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Stack;
 
-import de.alpharogroup.regex.RegExExtensions;
+import io.github.astrapi69.regex.RegExExtensions;
 import io.github.astrapi69.io.file.filter.MultiplyExtensionsFileFilter;
 
 /**