+++ src/test/java/io/github/astrapi69/csv/CsvFileExtensionsTest.java
@@ -42,12 +42,12 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.collections.CollectionExtensions;
-import de.alpharogroup.collections.array.ArrayFactory;
-import de.alpharogroup.collections.list.ListExtensions;
-import de.alpharogroup.collections.list.ListFactory;
-import de.alpharogroup.collections.map.MapFactory;
-import de.alpharogroup.collections.set.SetFactory;
+import io.github.astrapi69.collections.CollectionExtensions;
+import io.github.astrapi69.collections.array.ArrayFactory;
+import io.github.astrapi69.collections.list.ListExtensions;
+import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.collections.map.MapFactory;
+import io.github.astrapi69.collections.set.SetFactory;
 import io.github.astrapi69.io.StreamExtensions;
 import io.github.astrapi69.FileExtensions;
 import io.github.astrapi69.create.FileCreationState;