+++ src/test/java/io/github/astrapi69/read/ReadFileExtensionsTest.java
@@ -45,7 +45,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.collections.array.ArrayFactory;
+import io.github.astrapi69.collections.array.ArrayFactory;
 import io.github.astrapi69.io.StreamExtensions;
 import io.github.astrapi69.FileTestCase;
 import io.github.astrapi69.delete.DeleteFileExtensions;