+++ src/test/java/io/github/astrapi69/zip/ZipperTest.java
@@ -39,10 +39,10 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.collections.list.ListFactory;
+import io.github.astrapi69.collections.list.ListFactory;
 import io.github.astrapi69.io.file.namefilter.MultiplyExtensionsFilenameFilter;
 import io.github.astrapi69.io.file.namefilter.SimpleFilenameFilter;
-import de.alpharogroup.meanbean.factories.FileFactory;
+import io.github.astrapi69.meanbean.factories.FileFactory;
 import io.github.astrapi69.search.FileSearchExtensions;
 import io.github.astrapi69.write.WriteFileExtensions;